+++ CHANGELOG.md
@@ -6,7 +6,9 @@
 - Added new “Element is being saved for the first time” and “Element is a provisional draft” webhook filters.
 - Some webhook filters now disable other mutually-exclusive filters when active and enabled.
 - Pending webhook requests now have a “Send now” button, for manually triggering a webhook if the queue job was lost or failed. ([#35](https://github.com/craftcms/webhooks/issues/35))
+- Old webhook request activity is now purged during garbage collection. ([#57](https://github.com/craftcms/webhooks/discussions/57))
 - Added the “Initial Delay” setting. ([#53](https://github.com/craftcms/webhooks/discussions/53))
+- Added the “Purge Duration” setting.
 - Added `craft\webhooks\filters\ExclusiveFilterInterface`.
 
 ### Changed